+++ webpack.mix.js
@@ -1,6 +1,6 @@
 let mix = require('laravel-mix');
 
 mix.js(
-    'src/assetbundles/cloudflare/src/js/settings.js',
-    'src/assetbundles/cloudflare/dist/js/settings.js'
+    'src/assetbundles/cloudflare/src/js/cp.js',
+    'src/assetbundles/cloudflare/dist/js/cp.js'
 );